refactor: move Sheets client & services around
